+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [3.1.0](https://github.com/phalcon/cphalcon/releases/tag/v3.1.0) (2016-XX-XX)``
 - Added `Phalcon\Validation\Validator\Callback`, `Phalcon\Validation::getData`
+- Added the ability to truncate database tables
 
 # [3.0.3](https://github.com/phalcon/cphalcon/releases/tag/v3.0.3) (201X-XX-XX)
 - Fixed implementation of Iterator interface in a `Phalcon\Forms\Form` that could cause a run-time warning